[Ui.Api, Ui.Core] TimeoutMinutes in step commands from int to double

+++ Biflow.Ui.Api/Models/Step/PackageStepDto.cs
@@ -3,7 +3,7 @@ namespace Biflow.Ui.Api.Models.Step;
 [PublicAPI]
 public sealed record PackageStepDto : StepDto
 {
-    public required int TimeoutMinutes { get; init; }
+    public required double TimeoutMinutes { get; init; }
     public required Guid ConnectionId { get; init; }
     public required string PackageFolderName { get; init; }
     public required string PackageProjectName { get; init; }
